Move Groovy test to separate module
