+++ assertj-tests/assertj-integration-tests/assertj-core-groovy/.editorconfig
@@ -1,0 +1,3 @@
+[*.groovy]
+indent_style = space
+indent_size = 2
